drivers: ramdisk: Use build scripts for more generic disk image embedding
Make use of CMake build system to provide a generic fallback for the case
where GNU extensions are not available. The generic fallback embeds disk
image binary data as array data, including the binary data converted into
hex character from at build time. When GNU extensions are available for
use, .incbin directive is still used, as it is much more efficient, with
respect to build time (especially as size of the binary data to be
embedded increases).

Signed-off-by: Irfan Ahmad <irfan.ahmad@siemens.com>